Added a common file for js
